refactor: scope getByBlobUri to the generic repo hydration concern (review #7)
getByBlobUri read by blob content regardless of the current pointer, yet lived
on BaseResourceRepository — inherited by every repository, including the
hand-rolled non-generic TargetStatsStorage, with nothing marking it as a
hydration-only primitive. Move it to GenericResourceRepository (where the four
relation/scope wrappers NodeLoader uses already delegate it) and document that
it is read-regardless-of-pointer and only safe for graph hydration, so repos
with GC/lifecycle semantics tied to pointer state don't inherit it. Behavior and
the getByBlobUri-then-getById fallback policy in NodeLoader.load are unchanged.
